moved imgui files into a cpp header and excluded to avoid errors

+++ AxtEngine/src/platform/windows/WinInput.cpp
@@ -25,7 +25,7 @@ namespace axt {
 		const auto window = static_cast<GLFWwindow*>(App::GetApp().GetWindow().GetNativeWindow());
 		double x, y;
 		glfwGetCursorPos(window, &x, &y);
-		return { x, y };
+		return { (float)x, (float)y };
 	}
 
 	float WinInput::GetMouseXI() {
